// src/interfaces/Booking.ts
export interface IntBooking {
    id: number
    photo: string
    name: string
    order: string
    checkin: string
    checkout: string
    typeroom: string
    numroom: number
    price: number
    request: string
    amenities: any[]
    //photos: Photo[]
    type: string
    description: string
    state: string
}